商品削除機能の実装修正

+++ app/controllers/items_controller.rb
@@ -35,8 +35,10 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    if @item.destroy
-      redirect_to action: :index
+    if current_user.id == @item.user_id 
+      if @item.destroy
+        redirect_to action: :index
+      end
     end
   end
